+++ JustPDF/views/MainView.swift
@@ -1,0 +1,15 @@
+import SwiftUI
+
+struct MainView: View {
+    let document: Document
+    @State private var state = DocumentState()
+    
+    var body: some View {
+        MinimalPDFView(pdf: document.pdf, state: state)
+            .ignoresSafeArea()
+            .focusedSceneValue(\.state, state)
+            .overlay {
+                HUDView(state: state)
+            }
+    }
+}
